+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Renamed `max_iteration_stop_cryterion` method in the Algorithm class to `max_iteration_stop_criterion`
   - Removed (previously deprecated) `very_verbose` parameter in `Algorithm`'s run method.
   - Removed (previously deprecated) `axpby` method in DataContainer.
+  - Deprecate use of integer compression in NEXUSDataWriter.
 
 * 22.2.0
   - BlockGeometry is iterable